refactor(digest): move to clarify site digest params
