+++ recoco/apps/communication/digests.py
@@ -313,33 +313,6 @@ def make_recommendations_digest(recommendations, user):
     return recommendation_digest
 
 
-def make_site_digest(site):
-    """Return site informations as a dict"""
-
-    data = {
-        "name": site.name,
-    }
-
-    site_config = SiteConfiguration.objects.get(site=site)
-
-    data.update(
-        {
-            "description": site_config.description or "",
-            "sender_name": site_config.sender_name or "",
-            # sender_email is the key name used on Brevo's side, painful to update
-            "sender_email": site_config.contact_form_recipient or "",
-            "legal_address": site_config.legal_address or "",
-            "main_topic": site_config.main_topic or "",
-            "legal_owner": site_config.legal_owner or "",
-        }
-    )
-
-    if site_config.email_logo:
-        data["site_logo"] = utils.build_absolute_url(site_config.email_logo.url)
-
-    return data
-
-
 def make_project_survey_digest_for_site(user, project, site):
     """Return survey information as a dict for a given project on a given site"""
