+++ recoco/apps/projects/views/__init__.py
@@ -29,7 +29,7 @@ from recoco.apps.communication import digests
 from recoco.apps.communication.api import send_email
 from recoco.apps.geomatics import models as geomatics_models
 from recoco.apps.geomatics.serializers import RegionSerializer
-from recoco.apps.home.models import AdvisorAccessRequest
+from recoco.apps.home.models import AdvisorAccessRequest, SiteConfiguration
 from recoco.apps.projects.models import ProjectCreationRequest
 from recoco.utils import (
     check_if_advisor,
@@ -157,6 +157,20 @@ def project_moderation_project_refuse(request: HttpRequest, project_id: int):
     return redirect(reverse("projects-moderation-list"))
 
 
+def make_site_digest(site):
+    """Return site informations as a dict"""
+    # only used for emails that specifically comment on site
+    # for site params sent in all emails, look at recoco.apps.communication.api.get_site_params
+
+    site_config = SiteConfiguration.objects.get(site=site)
+    data = {
+        "name": site.name,
+        "main_topic": site_config.main_topic or "",
+        "legal_owner": site_config.legal_owner or "",
+    }
+    return data
+
+
 @login_required
 def project_moderation_project_accept(request: HttpRequest, project_id: int):
     is_project_moderator_or_403(request.user, request.site)
@@ -218,10 +232,8 @@ def project_moderation_project_accept(request: HttpRequest, project_id: int):
 
                 params = {
                     "project": digests.make_project_digest(project, owner),
-                    "site": digests.make_site_digest(
-                        project.project_sites.origin().site
-                    ),
-                    "survey_site": digests.make_site_digest(request.site),
+                    "site": make_site_digest(project.project_sites.origin().site),
+                    "survey_site": make_site_digest(request.site),
                     "survey": digests.make_project_survey_digest_for_site(
                         owner,
                         project,
